DDNS: Request ipv4 address by default. (#167)
When ipv6 and ipv4 are available for the given connection by default
ipv4 should be preferred, as most users will probably be interested to
update their ipv4 addresses.

There are two ways to accomplish this, either use the ipv4 subdomain
'ipv4.icanhazip.com' or force curl to resolve to ipv4 address. The
recommended way described in the FAQs
(https://major.io/icanhazip-com-faq/) is to use the command-line
options, which this change implements.