Fix Semantic error Schemas with 'type: array', require a sibling 'items: ' (#186)
Semantic error at components.parameters.SortDelegationSignerRecords.schema
Schemas with 'type: array', require a sibling 'items: ' field
Jump to line 2135

Actually, we allow to use any of those enums, of combination of them separated by comma. But there's no way to define it in the schema, so I'm just setting it as enum as I think was originally the plan.