Create ifstated.conf
Hi!

The main purpose of this ifstated.conf is to keep DNSimple API queries to a minimum, instead of calling ddns.sh every 5min via cron. The ddns.sh is still needed as quickly stated on the comments. The way it's implemented now is that ifstated runs ddns.sh only if it detects that the ping tests have failed.

It's a dirty hack, I'm sure more functionality can be added if needed. For example detecting quick dynamic DNS changes enforced by the ISP. Maybe using dig on special domains and comparing the IP against ifconfig outputs.

Let me know if further changes are needed! Thanks!