Remove Transfer-Encoding from the Fixtures
While running some tests on dnsimple-go we noticed that Go 1.15 dropped
support for Transfer-Encoding: identity on `net/http` and this is
causing our build to break.

This commit removes all Transfer-Encoding from all our fixtures and
updates the `bin/download_fixture.sh` script to not save it on future
fixtures.