Issue certificates use-case (#332)
This PR introduces a short tutorial on how to issue LetsEncrypt certificates, via our API and how to use webhooks to track when the certificate is ready for installation. It does not cover the actual installation of the certificate as the approach can be highly varied, depending on the infrastructure and method of storing the certificates e.g. is it stored in Chef Vault, Hashicorp Vault, AWS ACM/KMS and etc.