fix: Update broken fixtures (#816)
Over the last few years we made several changes to fixtures, sometimes even "manual", and we never ported them to clients.

When I tried to pull the fixtures from here into the dnsimple-go client... it was a bloodbath.

I fixed all fixtures in the Go client repo, and I am porting the fixtures here.